Make Harfbuzz accept DFLT language tag.
In fonts (such as Roboto) that have only a DFLT language tag in their
OpenType layout tables, Harfbuzz was skipping those features because it
was only matching against the explicit language tag. This patch makes
Harfbuzz check the DFLT tag too, in case there's not an explicit match.

Behdad looked at this patch and said it looks fine.

Note that committing this without having positioning fully in place will
cause regressions, as kerning makes the advances as generated by
Harfbuzz inconsistent with the advances actually used in drawText.

Change-Id: I6ea2eb2a580d00410c05c748e2db577072c25534