Fix to make space width consistent in bidi
This is a fix for bug 7363916, where the inconsistent breaking of the
text into script runs was leading to a different measured width on paint
than on measure for a TextView, leading to the content being ellipsized.

This patch classifies the space as Latin, rather than the script of the
surrounding text, which means that it's always drawn from the Latin
font, and thus is measured consistently.

Change-Id: I07abfa2918698c93f55333f1e964cc24a18d3def